Fix(liked): Prevent crash on blank favorites page
The favorites page was blank due to a runtime crash when data in local storage was malformed. The `useQuery` hook would fail when it received invalid arguments, causing the component to fail to render.

This was caused by a `useLocalStorage` hook that was not robust enough to handle unexpected data formats. The hook was also duplicated in `App.tsx` and `src/app/liked/page.tsx`.

This commit addresses the issue by:
1. Creating a single, robust, shared `useLocalStorage` hook in `src/hooks/useLocalStorage.ts`. This new hook validates the data retrieved from local storage and falls back to the initial value if the data is corrupted, preventing the application from crashing.
2. Refactoring both `App.tsx` and `src/app/liked/page.tsx` to use this new shared hook, eliminating duplicated code and ensuring consistent, safe access to local storage across the application.
