+++ src/App.tsx
@@ -10,28 +10,7 @@ import { AIQuestionGenerator } from "./components/ai-question-generator/ai-quest
 import { Link } from "react-router-dom";
 import { useTheme } from "./hooks/useTheme";
 import { AnimatePresence } from "framer-motion";
-
-function useLocalStorage<T>(key: string, initialValue: T) {
-  const [storedValue, setStoredValue] = useState<T>(() => {
-    try {
-      const item = window.localStorage.getItem(key);
-      return item ? JSON.parse(item) : initialValue;
-    } catch (error) {
-      return initialValue;
-    }
-  });
-
-  const setValue = (value: T) => {
-    try {
-      setStoredValue(value);
-      window.localStorage.setItem(key, JSON.stringify(value));
-    } catch (error) {
-      console.error(error);
-    }
-  };
-
-  return [storedValue, setValue] as const;
-}
+import { useLocalStorage } from "./hooks/useLocalStorage";
 
 export default function App() {
   const { theme, setTheme } = useTheme();
