+++ src/hooks/useLocalStorage.ts
@@ -1,0 +1,39 @@
+import { useState } from "react";
+
+export function useLocalStorage<T>(key: string, initialValue: T) {
+  const [storedValue, setStoredValue] = useState<T>(() => {
+    if (typeof window === "undefined") {
+      return initialValue;
+    }
+    try {
+      const item = window.localStorage.getItem(key);
+      if (!item) {
+        return initialValue;
+      }
+      const value = JSON.parse(item);
+      // If initialValue is an array, ensure the stored value is also an array.
+      // This prevents crashes if the data in localStorage is corrupted.
+      if (Array.isArray(initialValue) && !Array.isArray(value)) {
+        return initialValue;
+      }
+      return value;
+    } catch (error) {
+      console.error(error);
+      return initialValue;
+    }
+  });
+
+  const setValue = (value: T | ((val: T) => T)) => {
+    try {
+      const valueToStore = value instanceof Function ? value(storedValue) : value;
+      setStoredValue(valueToStore);
+      if (typeof window !== "undefined") {
+        window.localStorage.setItem(key, JSON.stringify(valueToStore));
+      }
+    } catch (error) {
+      console.error(error);
+    }
+  };
+
+  return [storedValue, setValue] as const;
+}
